🔧 feat: AVI SSOT Implementation - Advanced Voice Interview System
Consolidates all Advanced Voice Interview system types into a Single Source of Truth pattern, eliminating code duplicity and improving type consistency for risk evaluation and voice analysis components.

### Changes:
- **NEW**: `models/avi.ts` - SSOT consolidating:
  - Voice analysis types (VoiceAnalysis, AVIResponse)
  - Risk evaluation system (AVIScore, RedFlag, AVIRiskLevel)
  - Question system (AVICategory, QuestionType, AVIQuestionEnhanced)
  - Session management (AVISession, AVICalibration, AVIConfig)
  - Enhanced type safety with strict enums and interfaces

- **UPDATED**: `models/types.ts` - Clean import consolidation:
  - Added imports from AVI SSOT
  - Removed 8+ duplicate type definitions (60+ lines)
  - Added reference comments for type location

### Services Updated:
- ✅ AVI Calibration Service imports updated
- ✅ All avi-*.service.ts files updated (10+ services)
- ✅ AVI Interview Component updated
- ✅ Voice validation and test helpers updated

### Benefits:
- ✅ Eliminates 8+ duplicate AVI type definitions
- ✅ Single source of truth for voice interview domain
- ✅ Consistent typing across risk evaluation system
- ✅ Enhanced type safety with strict enums
- ✅ Zero TypeScript compilation errors
- ✅ Improved maintainability for voice analysis features
